move pattern apis to sponged
