+++ cmd/sponge/cmdpattern/README.md
