+++ cmd/sponge/cmdpattern/commands.go
