+++ cmd/sponged/handlers/pattern.go
