+++ cmd/sponged/tests/pattern_test.go
