+++ config/localhost.cfg
@@ -46,7 +46,7 @@ export WIRE_HOST=:16182
 ##############
 # XENIAD
 
-export XENIA_MONGO_HOST=mongodb://localhost:27017/coral
+export XENIA_MONGO_URI=mongodb://localhost:27017/coral
 export XENIA_AUTH_PUBLIC_KEY=
 
 export XENIA_LOGGING_LEVEL=1
